Split out test_email
